+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.6: 2024-10-20
 
 * Fix regression: Completed tasks being ignored when syncing calendar events to Todoist
+* Capitalize the first letter in the Finnish weekday on note header
 
 # 1.1.5: 2024-10-19